Specify desired read.table column types (for R 4.x compatibility)
In R 3.x, character columns default to factors; in R 4.x, they default
to character vectors. Enable this script to work with all R versions
by being explicit for `static`: we want a character vector for Chrom
(column 1), but a factor for Feature (column 5).

The previous code running in R 3.x had Chrom as a factor too.
To print each myFeatureChr[[]] entry, it reduced it to a single level
(via droplevels()) and printed that single level rather than its value.
Forcing this column to be a character vector instead means this code
can be simplified.

For `data`, Chrom (column 1) changes from factor to vector too,
but the code works with either so we don't override the default.

Bump geneCN.R version number to 2.2.